Move matrix wrappers to Hamiltonian/ directory
This adresses comments in pull request #18
