+++ CMakeLists.txt
@@ -58,13 +58,13 @@ endif()
 set (NETKET_SOURCES ${NETKET_SOURCE_DIR}/netket.cc)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Graph)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Hamiltonian)
+set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Hamiltonian/MatrixWrapper)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Headers)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Hilbert)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Learning)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Lookup)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Machine)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Observable)
-set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Observable/MatrixWrapper)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Sampler)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Stats)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Utils)
