+++ NetKet/Hamiltonian/MatrixWrapper/abstract_matrix_wrapper.hpp
