+++ NetKet/Hamiltonian/MatrixWrapper/dense_matrix_wrapper.hpp
