+++ NetKet/Hamiltonian/MatrixWrapper/direct_matrix_wrapper.hpp
