+++ NetKet/Hamiltonian/MatrixWrapper/sparse_matrix_wrapper.hpp
